+++ test/cli/test_main.py
@@ -291,7 +291,7 @@ class OperationCommandTest(CliTestCase):
         self.assert_main(['op', 'list', '--internal'], expected_stdout=['2 operations found'])
         self.assert_main(['op', 'list', '--tag', 'input'], expected_stdout=['9 operations found'])
         self.assert_main(['op', 'list', '--tag', 'output'], expected_stdout=['6 operations found'])
-        self.assert_main(['op', 'list', '--deprecated'], expected_stdout=['One operation found'])
+        self.assert_main(['op', 'list', '--deprecated'], expected_stdout=['2 operations found'])
 
 
 @unittest.skip(reason='Hardcoded values from remote service, contains outdated assumptions')